+++ fastly/ngwaf/v1/signals/api_list.go
@@ -1,0 +1,44 @@
+package signals
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strconv"
+
+	"github.com/fastly/go-fastly/v10/fastly"
+)
+
+// ListInput specifies the information needed for the List() function to perform
+// the operation.
+type ListInput struct {
+	// Context, if supplied, will be used as the Request's context.
+	Context *context.Context
+	// Limit how many results are returned.
+	Limit *int
+	// WorkspaceID is the workspace identifier (required).
+	WorkspaceID *string
+}
+
+// List retrieves a list of signals for the given workspace, with optional filtering and pagination.
+func List(c *fastly.Client, i *ListInput) (*Signals, error) {
+	requestOptions := fastly.CreateRequestOptions(i.Context)
+	if i.Limit != nil {
+		requestOptions.Params["limit"] = strconv.Itoa(*i.Limit)
+	}
+
+	path := fastly.ToSafeURL("ngwaf", "v1", "workspaces", *i.WorkspaceID, "signals")
+
+	resp, err := c.Get(path, requestOptions)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	var signals *Signals
+	if err := json.NewDecoder(resp.Body).Decode(&signals); err != nil {
+		return nil, fmt.Errorf("failed to decode json response: %w", err)
+	}
+
+	return signals, nil
+}